pyproject.toml
[project]
name = "opensource-clipping"
version = "0.9.5"
description = "AI Auto-Clipper & Teaser Generator — converts long videos into viral short-form clips"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "yt-dlp[default]",
    "faster-whisper",
    "google-genai",
    "openai",
    "mediapipe",
    "opencv-python",
    "Pillow",
    "requests",
    "python-dotenv",
    "google-api-python-client",
    "google-auth",
    "google-auth-oauthlib",
    "google-auth-httplib2",
    "ultralytics",
    "gdown",
]

[project.scripts]
clipping = "main:main"
